[Refactor]: Improved support for runtime theme update, mod+arrow handling. lib/cluterm:   - removed the global initial configs.   - inital configs are now stored in cluterm, instead of global singleton.   - remove 'cell' module, became unnecessary. frontend:   - theme now included in snapshot for rendering.   - osc query format update to use 'rgb:RRRR/GGGG/BBBB'.   - Added support for mod+arrow keys (tmux pane resize and stuff). Some code refactor, cleanup and stuff.

+++ lib/cluterm/vt/parser.h
@@ -75,7 +75,7 @@ typedef struct VT_Parser {
 } VT_Parser;
 
 void parser_init(VT_Parser *);
-void parser_feed(VT_Parser *, const uchar *, uint32_t);
+void parser_feed(VT_Parser *, const uchar *, size_t);
 FSM_Event parser_run(VT_Parser *);
 
 #endif
